refactor: move components to another directory
