+++ src/components/NotePreview/NotePreview.tsx
